+++ backends/k8s/cogserver-dev/scripts/uninstall.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-kubectl delete -f yaml/titiler-ingress.yaml
-kubectl delete -f yaml/titiler-service.yaml
-kubectl delete -f yaml/titiler-deployment.yaml
-kubectl delete -f yaml/titiler-namespace.yaml