+++ packages/svelte-undp-design/src/lib/CardWithImage/CardWithImage.svelte
@@ -8,6 +8,7 @@
 	export let title: string;
 	export let image: string;
 	export let accent: 'global' | 'yellow' | 'red' | 'green' | 'blue' = 'global';
+	export let icon = '';
 
 	export let width: number;
 	export let height: number;
@@ -16,13 +17,20 @@
 </script>
 
 <div
-	class="content-card {accent === 'global' ? '' : `accent-${accent}`} {tag
+	class="content-card {accent === 'global' ? '' : `accent-${accent}`} {tag || icon
 		? ''
 		: 'hide-border-top'}"
 >
 	<a href={url}>
-		{#if tag}
-			<h6 class="" data-viewport="false">{tag}</h6>
+		{#if tag || icon}
+			<h6 class="" data-viewport="false">
+				{#if icon}
+					<i class={icon}></i>
+				{/if}
+				{#if tag}
+					{tag}
+				{/if}
+			</h6>
 		{/if}
 		<div class="image">
 			<img