+++ sites/geohub/src/lib/config/DefaultUserConfig/MapPageTableViewType.ts
@@ -1,0 +1,3 @@
+import type { TableViewType } from '$lib/types';
+
+export const MapPageTableViewType: TableViewType = 'card';