+++ packages/svelte-undp-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @undp-data/svelte-undp-components
 
+## 0.1.2
+
+### Patch Changes
+
+- bc616e0: fix: close ModalTemplate by espace key if close button is shown in dialog.
+
+## 0.1.1
+
+### Patch Changes
+
+- 6b7907a: fix: add fontWeight prop to FieldControl
+
 ## 0.1.0
 
 ### Minor Changes