+++ packages/svelte-undp-design/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @undp-data/svelte-undp-design
 
+## 1.0.7
+
+### Patch Changes
+
+- 6255c69: fix: add icon option to show fa icon next to tag at Card and CardWithImage component
+
+## 1.0.6
+
+### Patch Changes
+
+- 5566f74: fix: add an option of forbiddenCharacters in SearchExpand component. as default, it ignore & and | char.
+
 ## 1.0.5
 
 ### Patch Changes